Template-ify trackless::Michels (#79)
* Template trackless michels class.

This both solves the include problem I was having AND it prepares this
class for MAT.

Also: neuter Cluster. It doesn't need CVU dependence.
