+++ includes/Cuts.cxx
@@ -38,9 +38,32 @@ A not-brief note on how the "exclusive" pion cuts work:
 #include "Cuts.h"
 
 #include "CutUtils.h"  // GetHadIdxsFromMichels, IsPrecut, GetWSidebandCuts, kCutsVector
-#include "Michel.h"  // endpoint::Michel, endpoint::MichelMap, endpoint::GetQualityMichels, trackless::GetQualityMichels
+#include "Michel.h"  // endpoint::Michel, endpoint::MichelMap, endpoint::GetQualityMichels
+#include "MichelTrackless.h"  //  trackless::GetQualityMichels
 #include "TruthCategories/Sidebands.h"  // sidebands::kSidebandCutVal
 #include "utilities.h"                  // ContainerEraser
+
+//==============================================================================
+// UTILITY
+//==============================================================================
+// Get pion candidate indexes from michel map
+// (our cuts strategy enforces a 1-1 michel-pion candidate match)
+std::vector<int> GetHadIdxsFromMichels(
+    const endpoint::MichelMap endpoint_michels,
+    const trackless::MichelEvent<CVUniverse> vtx_michels = trackless::MichelEvent<CVUniverse>()) {
+  std::vector<int> ret;
+
+  // endpoint michels
+  for (auto m : endpoint_michels) ret.push_back(m.second.had_idx);
+
+  // vertex michels
+  // When m_idx is set (i.e. != -1), then we have a good vertex michel.
+  // In that case, a -1 in this hadron index return vector is the code that for
+  // this analysis that we have a good vertex michel.
+  if (vtx_michels.m_idx != -1) ret.push_back(-1);
+
+  return ret;
+}
 
 //==============================================================================
 // Passes ALL Cuts
@@ -59,7 +82,7 @@ PassesCutsInfo PassesCuts(CVUniverse& universe, const bool is_mc,
   // passes all cuts but w cut
   //============================================================================
   endpoint::MichelMap endpoint_michels;
-  trackless::MichelEvent vtx_michels;
+  trackless::MichelEvent<CVUniverse> vtx_michels;
   bool passes_all_cuts_except_w = true;
   for (auto c : GetWSidebandCuts()) {
     // Set the pion candidates to the universe. The values set in early cuts
@@ -137,13 +160,13 @@ EventCount PassedCuts(const CVUniverse& univ,
 // Pass Single, Given Cut v2
 // NEW
 // passes_this_cut, endpoint_michels, vtx_michels
-std::tuple<bool, endpoint::MichelMap, trackless::MichelEvent> PassesCut(
+std::tuple<bool, endpoint::MichelMap, trackless::MichelEvent<CVUniverse>> PassesCut(
     const CVUniverse& univ, const ECuts cut, const bool is_mc,
     const SignalDefinition signal_definition, const endpoint::MichelMap& em,
-    const trackless::MichelEvent& vm) {
+    const trackless::MichelEvent<CVUniverse>& vm) {
   bool pass = false;
   endpoint::MichelMap endpoint_michels = em;
-  trackless::MichelEvent vtx_michels = vm;
+  trackless::MichelEvent<CVUniverse> vtx_michels = vm;
   const bool useOVMichels = false;
 
   if (IsPrecut(cut) && !is_mc) return {true, endpoint_michels, vtx_michels};
@@ -394,6 +417,9 @@ bool PmuCut(const CVUniverse& univ) {
          pmu < CCNuPionIncConsts::kPmuMaxCutVal;
 }
 
+
+
+
 //==============================================================================
 // BEING DEPRECATED
 //==============================================================================
@@ -529,7 +555,7 @@ bool PassesCut(const CVUniverse& univ, const ECuts cut, const bool is_mc,
         else
           endpoint_michels.insert(m);
       }
-      trackless::MichelEvent mehreen_michels =
+      trackless::MichelEvent<CVUniverse> mehreen_michels =
           trackless::GetQualityMichels(univ);
       return endpoint_michels.size() > 0;  // || mehreen_michels.size() = 0;
     }
